requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
aiohttp>=3.10.11
aiofiles>=23.2.0
cryptography>=3.4.8
paho-mqtt>=2.1.0
python-dotenv>=1.2.1
aiosqlite>=0.20.0
aiohttp-socks>=0.9.0
